+++ CHANGELOG.md
@@ -9,7 +9,11 @@ All notable changes to this project will be documented in this file based on the
 
 ### Added
 
+* [Field](https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-function-score-query.html#function-random) param for `Elastica\Query\FunctionScore::addRandomScoreFunction`
+
 ### Improvements
+
+* Using `Elastica\Query\FunctionScore::addRandomScoreFunction` without `$field` parameter is deprecated since ES 6.0 and will fail since ES 7.0
 
 ### Deprecated